refactor: Move PropertyPathFilter to PropertyPathFilterDefinition
