--- library/ExternalContent/PropertyPathFilter/Contracts/Enums/Operator.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums;
-
-enum Operator
-{
-    case EQUALS;
-    case NOT_EQUALS;
-}
