--- library/ExternalContent/PropertyPathFilter/Contracts/Enums/Relation.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums;
-
-enum Relation
-{
-    case AND;
-    case OR;
-}
