+++ library/ExternalContent/PropertyPathFilterDefinition/Contracts/Enums/Relation.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums;
+
+enum Relation
+{
+    case AND;
+    case OR;
+}
