+++ library/ExternalContent/PropertyPathFilterDefinition/Contracts/RuleInterface.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilter\Contracts;
+namespace Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts;
 
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Operator;
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Operator;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Relation;
 
 interface RuleInterface
 {
