+++ library/ExternalContent/PropertyPathFilterDefinition/Filter.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilter;
+namespace Municipio\ExternalContent\PropertyPathFilterDefinition;
 
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\FilterInterface;
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\RuleSetInterface;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\FilterInterface;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\RuleSetInterface;
 
 /**
  * Class Filter
