+++ library/ExternalContent/PropertyPathFilterDefinition/Rule.test.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Municipio\ExternalContent\PropertyPathFilter;
+namespace Municipio\ExternalContent\PropertyPathFilterDefinition;
 
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Operator;
-use Municipio\ExternalContent\PropertyPathFilter\Contracts\Enums\Relation;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Operator;
+use Municipio\ExternalContent\PropertyPathFilterDefinition\Contracts\Enums\Relation;
 use PHPUnit\Framework\TestCase;
 
 class RuleTest extends TestCase
